Add files via upload
Task 1
This program shows what will happen when we dont include the header file "<iostream>" and run the program.

As the output clearly shows that cout function isnt declared. This is because the header files contain functions such as 
standard I/O functions and other functions too. There are different header files in different languages that have some functions which
make the program understand what to do.

Task 2
The code represents output when the header file <iostream> is presented as <iostream.h>

the output gives out error as no such file or directory, this is because "iostream.h" is deprecated and not a
standard header. It was used in older programs before C++ was standardized, Functions like cout were defined inside iostream.h

After C++ got standardized, all these functions like cout got moved into the std namespace. TO adjust to this change, non .h versions of some of these
headers were introduced and declared as standard headers.